css and layout for add product
