+++ src/components/NewProduct.css
@@ -1,0 +1,3 @@
+.productH1{
+    color: black;
+}
